tools/veyron: new version of the veyron 'update' command
The new version of the 'update' command scans the local file system to
identify what veyron "projects" exist and compares this view to the
view defined by the latest version of the project manifest. It then
determines what type of operation (create, delete, move, or update) is
needed to update the current state to the new state, and executes the
appropriate action. Notably, new projects are automatically configured
to for veyron's code review system by installing the appropriate git
hooks.

After this CL, adding new projects or moving and removing existing
projects will be as simple as updating the project manifest and
running 'veyron update'.

Change-Id: If34694e0633620f6a9d7d2aa98bcf56fa406243d