tools/veyron: adding support for ignoring local repositories during the update operation
The 'veyron project update' command scans the local file system to
identify git repositories under $VEYRON_ROOT so that it can determine
what operations are to be performed when updating the current state of
the local projects to the state described by a manifest.

If there are multiple copies of the same repositories, the update
operation is generally undefined and terminates with an error. This CL
adds support for a ".veyronignore" file that informs the update
operation which directories are to be ignored. Using this mechanism,
developers can indicate to the update tool, which local repositories
are to be excluded from the update operation.

Change-Id: I4d119437580bc276af5b1b035dd2d3196148c355