tools/presubmitter: add "test" command
The "test" command will pull changes from Gerrit into a new branch, run tests,
post the test results summary back to Gerrit, and clean up the branch.

The tests to run for each repo is specified in a config file (presubmit_test.xml).
In this CL, I created a proof of concept config file. If it looks good, I will
complete it.

Change-Id: I4ca643e90d5e8d62af693a8e6bc9beec709d00a1