tools/veyron: Make veyron go only generate Go from vdl files.
Since "veyron go" only builds Go source, it makes sense to limit
it to generating Go from vdl files.

Also changed "veyron go" and "veyron run" such that if the exec
command fails with a non-zero status, we simply call os.Exit with
that status.  The idea here is that "veyron go" and "veyron run"
should behave as if we were actually running the exec command
directly.  In practice it means that we get less spam on stderr
when things go wrong.

Change-Id: Ie4f872f0a1029534db9afe8916904e3f54ccef72