tools: cleanup, refactoring of running utility for tools, and initial tests for the veyron tool
The main change introduced by the CL is the replacement of the "cmd"
package for both running commands and logging their output with a
thinner "runutil" package. Refactoring of this logic was prompted by
writing tests for the veyron tool, which constitue the other part of
this CL. The set of tests for the veyron tool is to be extended, but I
would like to do a review now, before the CL gets too big.

Change-Id: I99c927bae025b2c4031e2ccbcb42199c730f00cf