tools/lib/cmdline: Copy latest cmdline from veyron repo.
This includes our latest changes that add ErrExitCode, as well as
better recursive help.

Also added package doc for every binary using the cmdline
package, and fixed some minor typos.  Note that the "long"
description of each root command should be of the form "The foo
tool ..." so that the generated godoc is idiomatic.

Change-Id: I9bcbc550dc12a3c36dc3d761c48903d6e51ff482