TBR: tools/presubmitter: change the "test" command to output test results to a test report file. Also change the "post" command to get review message from a file.
The main reason of this change is that I want to include which tests are failing
in the presubmitter review message. To get failed tests, I need to wait for
xUnit plugin to finish generating JUnit report (a post-build step), and use a
Jenkins API to get the failed tests after that. I can no longer post review in
the "build" phase because at that moment the JUnit report is not ready yet.

Here is the new process:
- At the "build" phase, run "presubmitter test", which will generate an
  intermediate test report file.
- At the beginning of the "post-build" phase, xUnit generates JUnit test results
  files (xml files).
- After that, in another "post-build" step, run "presubmitter post", which will
  read the intermediate test report file generated by "presubmitter test",
  use Jenkins API to get failed tests, and post the content of the intermediate
  test report file along with the failed tests to Gerrit.

Change-Id: Ie7930515ac409e0efa0b46a8c9bcba48f736cb42