tools/veyron: Make sure vdl update is run with correct env.
The previous logic was explicitly setting only the GOPATH
environment variable when running vdl.  That's brittle though;
e.g. switching to VDLPATH means we'd need more special-cases.

The new logic does a full SetupVeryronEnvironment for the host
platform before running the vdl tool, and then sets up the
environment for the go tool.

Change-Id: If534ec868750d1837cad187e6595eabf8aa94da6