tools/presubmitter: fix a report link generation bug.
When two test cases have the same class name and test name, we append "_2" to
the report link of the second one. In the code, the "seenTests" map is tracking
how many times a test has been seen. But before this CL, this map only tracks
the failed tests, not all the tests.

Change-Id: Id6c4dc09d11d2d39194b6a59fd58c986435e3a04