tools/lib/util: Remove SetupVeyronEnvironment.
Our previous strategy was to capture the base set of environment
variables at process startup, and always use that base when
generating a new veyron environment.  The purpose of this was to
avoid repeatedly adding values to vars like GOPATH, and was a
result of the fact that we modified the environment.

Capturing the environment at startup may be unexpected for
developers, and makes internal testing difficult, since we can't
modify the environment in the tests.

The new strategy is to pass the veyron environment explicitly
through to subcommands.  We never call os.Setenv to mutate the
environment.

To make this easier to understand, I've added envutil.Snapshot,
which helps manage snapshots of the environment.  I've also
looked through every call to exec.Command and runutil to ensure
we're passing the environment through everywhere we care.

Change-Id: I9a922ebfc3a5ca4534a487549c1b6eb901cd89dc