tools/veyron: Limit "veyron go" to generate vdl of affected pkgs.
We now limit vdl generation to the packages that are actually
specified on the command-line, rather than always running "vdl
generate all".

This is tricky and inexact; see comments in the implementation
for details.  Practically speaking this works pretty well; the
main thing we'd like to avoid is for "veyron go install foo" to
update completely unrelated vdl files, in other repos.

Change-Id: I1e0a4faf5f266170028f19f31573485e3f016788