tools/veyron: adding "veyron build" command
The build command can be used to manage veyron builds. In particular,
it can be used to list known builds and generate new builds.

The builds are represented as manifests and are revisioned using the
manifest repository located in $VEYRON_ROOT/.manifest. Each build is
identified with a tag, which the $VEYRON_ROOT/tools/conf/veyron
configuration file associates with a set of jenkins projects that
determine the stability of the build.

Internally, build manifests are currently organized as follows:

<manifest-dir>/
  builds/
    <tag1>/
      <tag1-build1>
      <tag1-build2>
      ...
    <tag2>/
      <tag2-build1>
      <tag2-build2>
      ...
    <tag3>/
    ...
  <tag1> # a symlink to a one of <tag1-build*>
  <tag2> # a symlink to a one of <tag2-build*>
  ...

Change-Id: I2e477fd8d2efba272052f3dbd995d978f00a9dfa