tools: porting implementation of the "veyron profile" logic from shell to go
This CL recreates the "veyron profile" logic in go. In addition, this
CL makes several improvements to the profile installation logic:

- the profile installation now uses a concept of atomic actions that
either succeed, or have no effect

- all commands executed in the course of a profile installation are
  now reported so that the user can better understand what went wrong

- when a profile installation command fails, the user is provided both
  the standard and the error output of the command that failed

In the course of this CL, I had to generalize the runtil.Run library,
which requires changes across all of our tools and also prompted me to
clean up some of our tests.

NOTE: There is room for refactoring in the profile.go, but I would
like to leave that as a TODO. Also, this CL does not come with any
tests, as testing all of the profile installations requires >30
minutes. Instead, I have tested this CL on a clean install of both
linux and darwin and all profiles installed successfully.

Change-Id: Idfb917c11d22fa173a57fc4c314d344ad235ac0a