tools: Make "veyron integration-test run" behave correctly when receiving "ctrl+c".
Before this CL, runutil.TimedCommand would create a new process group for the
command because Setpgid is set to true. This is problematic in some cases where:
- program A runs a TimedCommand B
- user presses "ctrl+c" for A, or A is killed by SIGTERM.
- A receives SIGINT/SIGTERM
- B cannot receive these signals because A and B are in different process groups.

This CL adds a flag to control whether to create a new process group for
TimedCommand. This is set to off in "veyron integration-test" command to make it
work correctly when receiving ctrl+c

I also:
- changed the code to use "go install" instead of "go run" to make running
  shelltest-runner simpler.
- made verboseFlag work better with the integration-test command.

Change-Id: Ib8679a7147b6f1e3e406a4ce4145f88fb976becc