tools/veyron/profile.go: Download and compile the go ppapi compiler as part of the "web" profile.
I decided that doing this in the web profile was better than adding the
compiler repo to the default manifest was better for a few reasons:
* It requires some dependencies (python-jinja2, and the nacl-sdk
  pepper_35 stuff)
* It has a build step, which only needs to run once.
* It's pretty big (takes 5 min to download) and only developers working
  on the js extension will need to actually build it.

This CL also removes the "build-nacl-compiler" test.  It has already
been removed from Jenkins.  That test was testing an old version of
compiler that we arn't using any more, and the current version is not
under active development any more anyways.

Lastly, the "build-extension" test has been updated to build the
extension in the new way.

I have a related CL that adds stuff to .gitignore and updates some
downloaded files in environment.

Change-Id: I3a468bfc5f511bac87e0088e5c73d356b1a0ebfb
MultiPart: 1/2