tools: Remove -novdl option from "veyron go" and "veyron xgo".
Also allow "veyron run go" to work, but with a warning (rather
than failing with an error).  There is a subtle difference
between "veyron go" and "veyron run go", where the former builds
VDL files and the latter doesn't.  But we need a simple way for
VDL developers to build the vdl tool, without updating all vdl
files.  Thus we allow "veyron run go", but have it emit a warning
message.

The -novdl option was meant as an easy way to make vdl codegen
changes, since we'd skip the "vdl generate" part.

But it has turned into a half-assed solution, since anything down
the chain that calls "veyron go" would forget about this setting,
and end up generating the vdl files with the old vdl compiler
anyways, leading to lots of confusion.

A better solution for vdl tool developers is to just always copy
the built vdl tool into $VEYRON_ROOT/bin, or to put
$VEYRON_ROOT/veyron/go/bin in their PATH before $VEYRON_ROOT/bin.
The latter may lead to weird problems if they run "veyron update"
but don't re-build the vdl tool, but we figure the vdl tool devs
are already pretty savvy, and can figure this out.

Put another way, the -novdl flag was no better than copying the
built vdl tool into $VEYRON_ROOT/bin, so we might as well make
vdl tool devs do that.

Change-Id: Ica850f6cdf14fb5a797c7c1422db9824d486b0ea