tools/testutil: add suffix to test case names to make them different in GoTest and GoRace.
The link generation for duplicated test cases is not consistent in Jenkins.
To mitigate this issue, we add a suffix to test case names to make them different in
the report from GoTest and GoRace test. This also make reading test report easier:
no more confusing duplicated test names like this:
https://dev.v.io/jenkins/job/veyron-presubmit-test/5271/testReport/veyron/io_veyron_veyron_runtimes_google_ipc/

Change-Id: I842e325c5db969d6a9d254de3baf25a78d69a4b4