tools/lib: clean up the way environment is managed by the testutil package and a couple of related changes:
1) envutil.LookPath() is now thread-safe

2) runutil.Run.Command...(file, args...) invokes envutil.LookPath() to
check if it can find a binary for <file> in the runutil.Run.Opts.Env
and if it succeeds it passes this binary to the underlying
exec.Command() call. In addition, runutil.Run.Command start a child
processes with a copy of its enviroment applying the contents of
runutil.Run.Opts.Env as a delta.

Change-Id: I935f5eeb021880aaa37b982b8fd3342150ceff90