tools: Change "vcloud run" to always cd into tmpdir.
This makes it easier when copying multiple run files, and you
want to run your own cmdline, since in the common case you won't
need to use the weird +TMPDIR notation.

Also added an -outdir flag to "vcloud run", which causes the
contents of each tmpdir to be copied back to the local machine.
This makes it easier to collect output from each remote node.

Change-Id: I414521fed6f935d3d1f7d78166092b91f2ac0379