veyron.io/tools: fix flaky test
Due to a bug in Go1.4 (see https://github.com/golang/go/issues/9384),
using an empty struct as the value in a map can cause a crash. This CL
works around the issue by using the boolean "true" as the map value.

The workaround will not be required after Go1.4.1.

Change-Id: Id8d51f4b78226780d1513d1978dfdd5071bcaab5