git-v23: fixing and speeding up the Go format check
This CL, makes sure that the Go format check "git v23" executes uses
the "gofmt" command that operates over files (as opposed to the "go
fmt" command that operates over packages). Additionally, this CL
checks all files using a single "gofmt" invocation (as opposed to
forking a process for each file to be checked).

Change-Id: Ib2d5f1241b841fe0123d71990b9e821fddc85ae8