tools/testutil|runutil: generate a xunit report to capture unhandled errors from tests instead of failing the whole presubmit test.
For context, visit https://dev.v.io/jenkins/job/vanadium-presubmit-test/231/consoleFull,
and search for "Running test "vanadium-prod-services-test"".

This test failed during the package building phase, which was not captured in any
xunit report file (which will be generated during the service checking phase).
This CL adds a sort of "catch all" report for this kind of failures.

Change-Id: I42fca2956b6538356485f15b02ea256dc5dbd9e4