tools: decoupling project names and remotes
This CL introduces a new format for in-memory and on-disk
representation of vanadium projects.

Previously, projects were represented using the following structure:

type Project struct {
  Name string
  Path string
  Revision string
  Protocol string
}

The "Name" field served double purpose. It represented the remote of
the project repository and it also served as a project identifier.

This turned out to be problematic when I changed our continuous
integration setup to use a GCE mirror; instances of VANADIUM_ROOT that
use the GCE as a mirror would wind up with project "names" that were
different from the project names other VANADIUM_ROOT instances
use. For example, the presubmit tasks running in GCE see local
project with names matching the mirror but need to figure out what
googlesource project to pull the patchset to test from. To fix this
problem, I have previously introduced the "Alias" hack (go/vcl/1230).

This CL fixes the issue the right way by decoupling the project name
and remote repository:

type Project struct {
  Name string
  Path string
  Remote string
  Revision string
  Protocol string
}

Further, to persists the project name (e.g. for the sake of creating
valid snapshots), this CL serializes the in-memory project
representation to the <projectDir>/.v23 directory during "v23 update".

The .v23 directories are now used in place of ".git" or ".hg"
directories for the purpose of identification of vanadium
projects. Consequently, the support for .v23ignore has been removed as
non-v23 git or hg projects won't have the .v23 directory and will thus
be ignored.

Change-Id: I18b81df805f053d68f2e1b130c470f5e6b321c00