tools: refactoring of the "vcloud node ..." commands
This CL moves the Jenkins-specific logic out of the vcloud tool,
creating a separate vjenkins tool. In addition, the setup script to
run on the nodes created through "vcloud node create" is now specified
as a command-line flag. Finally, the ssh authorization logic is moved
to a separate command and de-authorization functionality is added.

The net result of this CL is a more flexible "vcloud node ..." command
set that can be used for management of a variety of GCE nodes (such as
production instances running the device manager).

Change-Id: Id2520b8b21aab7a54a6a7dd6fb5bebd81c66c3c8