lib/gitutil: CommitAmend should allow empty commits.
Cos was running into an issue where git v23 review was failing with the
following error:

    ERROR: 'git commit --amend -m Copyright header added.

    You asked to amend the most recent commit, but doing so would make
    it empty. You can repeat your command with --allow-empty, or you can
    remove the commit entirely with "git reset HEAD^".

I think this has something to do with changing only the CL commit
message, and none of the rest of the CL.  I'm not sure why it doesn't
happen more frequently.

I think the right thing to do is to allow empty commits to be sent to
Gerrit.  If the code is in a weird state, at least it will be easy to
see the code and send another patch with fixes.

Change-Id: I16afd6ec38b25ec460bb030d7c5de71757775f84