devtools/lib/testutil: fix a bug in generateDummyTestReportFile.
We use "v23 test run vanadium-presubmit-test" to run a presubmit test.
It calls the vanadiumPresubmitTest function in lib/testutil/presubmit.go.
The "test run" command will fail when
1. vanadiumPresubmitTest returns a non-nil error, which means the test logic
   itself failed, or
2. vanadiumPresubmitTest returns a test result that is not TestPassed.

In presubmit, we only want to fail it when "1" happens. (Actually now "1" will
not even fail the command because we wrap it in an internal test error and
write it to a xunit file). If there is no errors in test logic itself,
vanadiumPresubmitTest should always return TestPassed as the test result.

But, the current implementations of generateDummyTestReportFile sometimes
returns TestFailed, which is not correct. I changed it so that it only returns
report generation related errors. If no errors, vanadiumPresubmitTest will
then return TestPassed as the final step.

Change-Id: I11771c94d967182b512eedc89021b6e24ee345e2