devtools: move (and rename) generateDummyTestReportFile from presubmit.go to run.go
Before this CL, the presubmit xUnit file uploaded to gcloud was not processed
through generateDummyTestReportFile where we do several sanity checks to make
sure the xUnit file exists and valid. The "presubmit result" command relies
on these checks to work correctly. To make sure we always upload valid xUnit
files, I moved generateDummyTestReportFile to run.go right before we upload
it to gcloud, and renamed it checkTestReportFile.

The only "side effect" is that the non-presubmit test runs will also go through
this function, which might not be a bad thing. But I need to modify some tests
to make sure the base dir of the generated test report file exists.

Change-Id: I7b5460cb88bf7468f30612863d2b8d4141665d29