devtools/presubmit: add an option to hide OS value in summary
This is to hide [linux] from the following.

✔ ➔ ✔: vanadium-go-race [linux]

It is a multi-configuration job so it can work with multi-parts. An OS axis
is also configured that has only one value "linux" in order to constrain
where its sub-builds run. In this case, we don't really want to show "[linux]"
in its summary line.

Change-Id: I3268c0b46ad5f6a93e7676bcd6196c923ca05b65