Counterpart to https://vanadium-review.googlesource.com/#/c/9040/
Since the default behvaior of the vrpc tool will not be to not
talk to unrecognized servers, the --insecure flag is needed to
make it do so.

Ideally, the monitoring tests would be setup with a principal that
recognizes these services. That is left as a future excercise to
the author or his colleagues.

Change-Id: Iba0abe1718d5a5a3dde842f06ef605a4674d6643