v23: add v23 api command
This CL adds a v23 api command and two subcommands:

check: compares the public API of the any changed Go packages in the
user's working directory with the 'canonical' version stored in a .api
file. If there is a difference, the user is notified and a diff is
produced.

fix: updates the .api file in any directory containing changes. The
user may then add the .api file to their changelist and submit it.

Change-Id: I05399490da340d0a0d050b5f9fc8057f04dcf031