devtools/oncall: create data collector and server.
These two things are kind of related so I put them in a single CL.

"collector" talkes to
- GCM to collect service latency data
- GCE (using gsutil command) to collect GCE instance data (status for now)

I decided to keep a history of data files so that we can support "going back in
time" feature in the dashboard, which is cool. I've set a lifecycle config on
the bucket so that all files older than 180 days are automatically deleted.

"server" serves the latest data file as JSON. It also accepts a "ts" parameter
to specify a specific data file (timestamp) to load.

Some logic is similar to devtools/dashboard, but I think it still makes sense
to separate them because this one will be used internally only.

Change-Id: I5d9791e349fcc9f77767d718a91377b740741564