devtools/regtest: inception of the Vanadium regression test
This CL implements initial version of the vanadium-regression-test
Jenkins target. The target invokes Vanadium integration test using a
combination of Vanadium binaries from different daily builds.

This CL also remove the vanadium-go-doc target, which is no longer
relevant.

NOTE: Currently, the Google Storage bucket in which we store the
Vanadium binaries is not world-readable and one has to point the
V23_KEY_FILE environment variable to the GCE service account
credentials in order to run the regression test.

Change-Id: I849617596faa7d655f5af14e45736ef52657950a