devtools/v23: relaxing the copyright check
During a test of the external contributor workflow, it occurred to me
that the current copyright check requires an exact match for the
AUTHORS and CONTRIBUTORS files. These files exist in each repository
and I expect their content to diverge over time. Even if it does not,
the current copyright check requires that all public projects
contain the identical AUTHORS and CONTRIBUTORS files, which seems like
an overkill since each modification of the AUTHORS / CONTRIBUTORS list
would require a multi-part CL for all of our public projects. This
CL relaxes the check to only check for the expected header.

Change-Id: Ie770714742f7c740c365250a932aab77d8d69016