x/devtools: Add config for crosstool-ng-1.20.0, add version to old config
Add a config file for crosstool-ng-1.20.0. The config is as similar as
possible to that for crosstool-ng-1.19.0, but an important change is
that glibc is now used instead of eglibc. (Crosstool dropped support
for eglibc, which is no longer maintained, and the flaky eglibc svn
repository was causing problems in arm profile setup.)

For now, configs for both crosstool versions are checked in, and
the v23 profile still uses crosstool 1.19.0. A later CL will switch
this to 1.20.0

Change-Id: Ibb910f5df8e7deb04953ef9d35ccc18e3e7e7920