devtools/v23/internal/test: add test for create-instance.sh script.
The reason I had to modify prod service test a little bit is that we cannot
access https://xxx:xxx/auth/blessingsroot because the site doesn't have valid
ssl certificate, so we cannot generate a valid credentials on the fly. In this
case we have to use "--insecure" when invoking "vrpc signature".

Change-Id: I80a18baf0d7fa8eea3db7c2b5d10623b04c0602d