profile/syncbase: build leveldb/snappy for different platform
Currently leveldb/snappy are built for one platform only,
but there should be a mechanism to build them for both linux_amd64
and linux_386 for integration testing.

This approach puts leveldb/snappy libraries to
third_party/cout/GOOS_GOARCH/..., making it possible to build
libraries for different platforms on one machine.
GOOS and GOARCH variables are derived from v23 GOOS and GOARCH,
so in order to work correctly the syncbase code should be targeted
to platform used to build the v23 tool.

Change-Id: Ia77bffeea4c31502c704e96de3af0f3e103adc69