devtools/v23: refinement of the v23 [api|copyright] logic
If the check is invoked in a project that does not exist in the
project manifest, the v23 command will now issue a warning instead of
failing, avoiding the need for using -check-[api|copyright]=false for
projects that are not in the project manifest.

In addition, this CL creates a whitelist of projects for which the
copyright check is to be used (similar to the whitelist of projects
for which the api check is to be used).

Fixes: https://github.com/vanadium/issues/issues/544

Change-Id: I89420d8d2f2dd49517c829f304cb495bad141ab6