devtools/v23: decouple web and nodejs profiles
Prior to this CL the web profile would be a superset of the nodejs
profile. This is not desirable in the presence of the support for
updating and uninstalling profiles (as there is no mechanism for
recording such relationship between profiles); in particular, "install
nodejs", "install web", "uninstall web" would leave the
$V23_ROOT/.v23_profiles metadata thinking that the nodejs profile is
installed, while its artifacts would have been wiped clean by the
"uninstall web" operation.

To address this problem, this CL decouples the web and nodejs
profiles, renaming the web profile to nacl to better capture its
nature. The other parts of this multipart CL take care of replacing
occurences with "profile install web" with "profile install nodejs" or
"profile install nacl nodejs" as appropriate.

MultiPart: 6/6
Change-Id: I9eef71d21fca409ce2379fcd560da207ad981345