devtools/v23: adding a command-line flag to "v23 cl" for specifying the remote branch the review is for
In addition, this CL refactors the support for CL reviews, moving the
struct for recording review options from the v23 tool to the gerrit
package. The reason for this is to avoid making further changes to the
public API of the gerrit package (e.g. Push() and Reference()) when
new review options are introduction.

Change-Id: I34506d43b1e3ffed8e82b8228ec3a7ef1245cf41