devtools/v23: adding logic to cleanup dependent CLs
When a dependency of a dependent CL is submitted, the "v23 cl mail"
needs to find out about this to avoid attempting to update a submitted
CL and getting an error from Gerrit.

There is a couple of way to do this:

1) "v23 cl mail" scans the git log of the remote branch the operation
pertains to for the Change-Id identifier for each CL in the sequence
of dependent CLs leading to the current branch and if found, it assume
this CL has been submitted and skips sending it to Gerrit. The
downside of this approach is that the worst case time complexity is
linear in the size of the git log.

2) Developers are expected to cleanup their dependent CLs using "v23
cl cleanup <name>", which takes care of updating the metadata used to
record the dependency relationship to reflect the fact <name> has been
submitted. The downside of this approach is that requires an action on
the developer part.

This CL implements 2).

Change-Id: I043ea347d829c93e1d9a60b95e96b2ad43b00b49