devtools/v23: fixing topics for sequence of dependent CLs
Prior to this CL, the "v23 cl mail" command would internally use the
"git push origin HEAD:refs/for/master?topic=<foo>" mechanism to set
the CL topic. However, this mechanism sets the topic for all CLs in
the sequence of dependent CLs and based on my communication with the
Gerrit team, there is no way to specify different topics through the
above mechanism.

This CL replaces the above mechanism with one that relies on invoking
the REST API to set the topic only for the CL corresponding to the
branch from which the "v23 cl mail" command was invoked.

Change-Id: I68590df730a2aa6f71af5e997e6c752961a3fcc2