devtools/v23: moving the godepcop check from client-side to server-side
This CL is the first of several CLs in which I plan to move the
client-side checks currently performed by the "v23 cl mail" command to
server-side (that is have them performed by Jenkins instead). There
are several benefits to doing this:

1) The client-side check is optional and if developers skip it, they
can screw others over.

2) The client-side check has been getting increasinly slow, while the
server-side check will run in parallel with other presubmit tests.

3) The Jenkins target used for the presubmit check will be also used
for a postsubmit check, providing us with a signal of whether the
submitted code meets the dependency requirements.

Change-Id: I35d1a0a94de60395162fd664cac7573b79769df0