devtools/internal/runutil: add wrappers for os.Stat and os.Open
This CL adds runutil wrappers for os.Stat and os.Open and replaces
existing invocations of os.Stat and os.Open with these wrappers. This
makes it possible to track these invocations in the "v23 -v" output.

In addition, this CL performs minor cleanup of things that I stumbled
across when implementing this CL.

Change-Id: Ie67422fe73a0c98a5fdd5a639298c044398c5628